refactor: split search module for easier testing
